backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
httpx
python-dotenv
huggingface_hub
